refactor: [deps] убрал большинство цикличных зависимостей в компонентах
