+++ package.json
@@ -62,6 +62,7 @@
     "babel-loader": "^8.2.2",
     "babel-plugin-react-scoped-css": "^1.1.1",
     "chromatic": "^5.10.2",
+    "circular-dependency-plugin": "^5.2.2",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "cp-cli": "^2.0.0",
     "css-vars-ponyfill": "^2.4.5",
@@ -93,8 +94,8 @@
     "swiper": "^6.8.2",
     "testing-library-selector": "^0.2.1",
     "ts-jest": "^27.0.3",
-    "typescript": "^4.4.4",
     "ts-node": "^10.4.0",
+    "typescript": "^4.4.4",
     "url-loader": "^4.1.1",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.1",
