+++ src/components/popups/FindEntitiesUser/FindEntitiesUser.tsx
@@ -2,9 +2,11 @@ import './FindEntitiesUser.scss';
 
 import React, { useState } from 'react';
 
-import {
-  Checkbox, Avatar, Tooltip, Structure, Toast
-} from '../../../index';
+import Checkbox from '../../atoms/Checkbox';
+import Avatar from '../../atoms/Avatar';
+import Structure from '../../molecules/Structure';
+import Toast from '../../atoms/Toast';
+import Tooltip from '../../atoms/Tooltip';
 import { Info, Copy } from '../../../indexIcon';
 
 import { IUser } from '../../../types/projects.types';
