+++ webpack-lib.config.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require('path');
-
+const CircularDependencyPlugin = require('circular-dependency-plugin')
 // =========================================================================
 module.exports = {
   entry: {index:'./src/index.ts',indexIcon:'./src/indexIcon.ts'},
@@ -82,5 +82,16 @@ module.exports = {
   },
   plugins: [
     // new BundleAnalyzerPlugin(),
+    new CircularDependencyPlugin({
+      // exclude detection of files based on a RegExp
+      exclude: /node_modules/,
+      // add errors to webpack instead of warnings
+      failOnError: false,
+      // allow import cycles that include an asyncronous import,
+      // e.g. via import(/* webpackMode: "weak" */ './file.js')
+      allowAsyncCycles: false,
+      // set the current working directory for displaying module paths
+      cwd: process.cwd(),
+    })
   ]
 };
